Remove limitation for crashing soon after boot
If NetworkStack crashes soon after boot, this currently does not crash
the system, and keeps it in a bad state. Then the system will likely
crash later when Wifi or ConnectivityService cannot use NetworkStack.

This was originally added to avoid crashloops if writing the last crash
timestamp failed, but it has not proved useful, makes it difficult to
diagnose issues (crash logs are gone by the time Wifi crashes the
device), and keeps devices in a bad (no connectivity) state instead of
recovering immediately.

Bug: 282533638
Test: boot, kill com.android.networkstack.process, see reboot
Change-Id: I864b05b17faec5914329860fad14a821383d1906